requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-core
sentence-transformers
faiss-cpu
rank-bm25
pypdf
python-dotenv
groq
